Update 0.1.15
Huge update, this has been the biggest development so far. I created a fish spawning and deleting system. As the player moves about the water fish will spawn in and remove themselves. There are several known bugs with this system I intend to fix.

Bug list:
Fish can spawn in on top of each other.
Fish can spawn in on coordinates that are not whole numbers.
Equipment hitbox is not accurate to the equipment size.

I'm sure there are more.